+++ include/Components/CameraComponent.h
@@ -2,24 +2,10 @@
 
 class GameObject;
 
-class CameraComponent
+struct CameraComponent : public Component
 {
-public:
-	CameraComponent( Ogre::Camera* cam, Ogre::Viewport* view );
-
-private:
-	component_id id;
+	CameraComponent( GameObject& owner, Ogre::Camera* cam, Ogre::Viewport* view );
 
 	Ogre::Camera* camera;
 	Ogre::Viewport* viewport;
-
-	friend class Scene;
-};
-
-class Camera final : ComponentLocator
-{
-public:
-	Camera( Scene* scene = nullptr, component_id id = 0, size_t index = 0 );
-
-	void Destroy() override;
 };
