+++ include/Components/Component.h
@@ -1,21 +1,13 @@
 #pragma once
 
-#include <cstddef>
-#include <vector>
-#include <limits>
-
-class Scene;
-
-component_id GenerateUniqueComponentID();
-
-class ComponentLocator
+struct Component
 {
-public:
-	ComponentLocator( Scene* scene = nullptr, component_id id = 0, size_t index = 0 );
-	
-	virtual void Destroy() = 0;
-
-	Scene* scene;
 	component_id id;
-	size_t index;
+	GameObject owner;
+
+	Component( GameObject& owner );
+
+	static component_id GenerateUniqueComponentID();
 };
+
+bool operator==( const Component& first, const Component& second );
