+++ source/Components/BehaviorComponent.cpp
@@ -2,22 +2,7 @@
 #include "Components/BehaviorComponent.h"
 
 BehaviorComponent::BehaviorComponent( GameObject& owner, BehaviorFunction behavior ) :
-	behavior( behavior ),
-	id( GenerateUniqueComponentID() ),
-	owner( owner )
+	Component( owner ),
+	behavior( behavior )
 {
-}
-
-Behavior::Behavior( Scene& scene, component_id id, size_t index ) : ComponentLocator( &scene, id, index )
-{
-}
-
-void Behavior::Destroy()
-{
-	scene->MarkForDestroy( *this );
-}
-
-bool operator==( const Behavior& first, const Behavior& second )
-{
-	return first.id == second.id;
 }
