+++ source/Components/CameraComponent.cpp
@@ -1,17 +1,9 @@
 #include "Gunship.h"
 #include "Components/CameraComponent.h"
 
-CameraComponent::CameraComponent( Ogre::Camera* cam, Ogre::Viewport* view ) :
-	id( GenerateUniqueComponentID() ),
+CameraComponent::CameraComponent(  GameObject& owner, Ogre::Camera* cam, Ogre::Viewport* view ) :
+	Component( owner ),
 	camera( cam ),
 	viewport( view )
-{
-}
-
-Camera::Camera( Scene* scene, component_id id, size_t index ) : ComponentLocator( scene, id, index )
-{
-}
-
-void Camera::Destroy()
 {
 }
